% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EveryNth.R
\name{EveryNth}
\alias{EveryNth}
\title{Add empty values into a vector}
\usage{
EveryNth(x, nth, empty = TRUE, inverse = FALSE)
}
\arguments{
\item{x}{vector}

\item{nth}{specify which elements are removed, i.e., every nth}

\item{empty}{logical specifying where elements are dropped or replaced}

\item{inverse}{logical specifying whether to drop every nth (FALSE; default)
or save every nth (TRUE)}
}
\value{
Vector with every nth value removed or made into an empty string
}
\description{
This function is particularly useful function for placing tick marks between 
labels in ggplot objects
}
\details{
I found this function on Stack Overflow while looking for a way to 
add tick marks to ggplot plots. URL provided in references. Post with function 
code was created by user *adamdsmith*.
}
\references{
\url{https://stackoverflow.com/questions/34533472/insert-blanks-into-a-vector-for-e-g-minor-tick-labels-in-r}
}
