% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrontMatter.R
\name{FrontMatter}
\alias{FrontMatter}
\title{R Front Matter}
\usage{
FrontMatter(packages = NULL, rstudio = NULL)
}
\arguments{
\item{packages}{Vector of packages we want to display. Default is NULL saying
we don't want to return any package information. Otherwise it can
be "all" to display information for all packages loaded.}

\item{rstudio}{String specifying Rstudio version used. Default is NULL.}
}
\value{
A string with the front matter information. By default returns R
version and time.
}
\description{
This function is used to create a string of information regarding the
R/Rstudio version and R package details used for an analysis for
printing in documents. The returned string includes newline characters so
requires cat() to display properly.
}
\examples{
FrontMatter()
FrontMatter("ggplot2")
FrontMatter(packages = "all", rstudio = "1.1.423")

}
\author{
Richard T Gray, \email{Rgray@kirby.unsw.edu.au}
}
