% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLegend.R
\name{GetLegend}
\alias{GetLegend}
\title{Function to extract legend from a ggplot}
\usage{
GetLegend(myggplot)
}
\arguments{
\item{myggplot}{ggplot plot object}
}
\value{
Legend object as a separate plot object.
}
\description{
Function to extract the legend from a ggplot for manipulation in a
publication plot. The code was obtained from a forum referenced below.
}
\references{
\url{http://www.sthda.com/english/wiki/ggplot2-easy-way-to-mix-multiple-graphs-on-the-same-page-r-software-and-data-visualization#change-legend-position}
}
\author{
Richard T Gray, \email{Rgray@kirby.unsw.edu.au}
}
