% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadLibrary.R
\name{LoadLibrary}
\alias{LoadLibrary}
\title{A function to load packages}
\usage{
LoadLibrary(package)
}
\arguments{
\item{package}{character string (unquoted) specifying name of package}
}
\value{
none
}
\description{
This function is a wrapper for require() to check if a package is available 
and then attempts to install the package if it is unavailable. By rights 
should be called LoadPackage.
}
\examples{
LoadLibrary(tidyverse)

}
\references{
\url{https://stackoverflow.com/questions/5595512/what-is-the-difference-between-require-and-library}
}
\author{
Richard T. Gray, \email{Rgray@kirby.unsw.edu.au}
}
