% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageCitation.R
\name{PackageCitation}
\alias{PackageCitation}
\title{A function to get the citations of packages}
\usage{
PackageCitation(packages, bibtex = FALSE, file = NULL)
}
\arguments{
\item{packages}{Vector of package names we want to cite. If "all" display all
packages loaded.}

\item{bibtex}{If TRUE display bibtex citation only.}

\item{file}{If specified is a string giving the file path
and name.}
}
\value{
Returns a list of the citation information and outputs a bibtex file
if requested.
}
\description{
This function outputs the package citation information of a package
or all packages loaded in the current session. It is based on a post by 
Carl Boettinger referenced below.
}
\examples{
PackageCitation("all")

}
\references{
\url{http://www.carlboettiger.info/2012/03/20/citing-r-packages.html}
}
\author{
Richard T Gray, \email{Rgray@kirby.unsw.edu.au}
}
