% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotColors.R
\name{PlotColors}
\alias{PlotColors}
\title{Vectors of colors for my plots}
\usage{
PlotColors(set = c("main", "lines", "cbPalette", "hivcascade4", "hivcascade5",
  "crayons"))
}
\arguments{
\item{set}{Character string indicating a set of colors.}
}
\value{
Vector of character strings representing the chosen set of colors, in RGB.
}
\description{
Copyright (C) 2018 Karl Broman
GNU General Public License, version 3, available at https://www.r-project.org/Licenses/GPL-3
}
\details{
This function Creates different vectors of colors that I use for plots. It
was inspired by Karl Broman's brocolors.R function:
https://github.com/kbroman/pkg_primer/blob/gh-pages/example/stage5/R/brocolors.R
which he uses in his minimal R package example. I have copied a number of
pieces of that, particularly the crayon colors specifications.
}
\examples{
PlotColors()

}
\references{
\url{https://github.com/kbroman/pkg_primer/blob/gh-pages/example/stage5/R/brocolors.R}

\url{http://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors}
}
\seealso{
\code{\link{PlotCrayons}}
}
\author{
Richard T Gray, \email{Rgray@kirby.unsw.edu.au}
}
