% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotOptions.R
\name{PlotOptions}
\alias{PlotOptions}
\title{A function to specify ggplot options}
\usage{
PlotOptions(opt = c("main"))
}
\arguments{
\item{opt}{name of option to use in plot}
}
\value{
ggplot theme options object
}
\description{
This function stores all my favoured plotting options and themes for
ggplot plots
}
\examples{
PlotOptions()

}
\author{
Richard T. Gray, \email{Rgray@kirby.unsw.edu.au}
}
