% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveFigure.R
\name{SaveFigure}
\alias{SaveFigure}
\title{Wrapper function to save ggplots}
\usage{
SaveFigure(
  folder,
  filename,
  figure,
  format = ".png",
  width = 15,
  height = 10,
  units = "cm",
  ...
)
}
\arguments{
\item{folder}{Character string for the folder where figure will be saved.}

\item{filename}{Character string specifiying the file name.}

\item{figure}{ggplot object.}

\item{format}{String specifying the file format of the figure. Default ".png".}

\item{width}{Saved plot width. Default is 15 cm.}

\item{height}{Saved plot height. Default 10 cm.}

\item{units}{String specifying units for plot dimension. Default centimetres.}

\item{...}{Additional ggsave inputs.}
}
\value{
None. Saves file in specified directory.
}
\description{
This is ggsave Wrapper function with defaiults so we don't have to keep
entering everything in ggsave.
}
\author{
Richard T Gray, \email{Rgray@kirby.unsw.edu.au}
}
