% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectCrayons.R
\name{SelectCrayons}
\alias{SelectCrayons}
\title{Select crayon colors}
\usage{
SelectCrayons(color_names = NULL, ...)
}
\arguments{
\item{color_names}{Optional vector of color names; can be partial matches.}

\item{...}{Additional optional color names.}
}
\value{
Vector of named RGB colors.
}
\description{
Copyright (C) 2018 Karl Broman
GNU General Public License, version 3, available at https://www.r-project.org/Licenses/GPL-3
}
\details{
Vector of colors selected from Crayola crayons. It is essentially a copy 
(with some minor tweaks) from Karl Broman's brocolors.R function:
https://github.com/kbroman/pkg_primer/blob/gh-pages/example/stage5/R/brocolors.R
}
\examples{
SelectCrayons("purple m")
SelectCrayons(c("purple m", "tickle"))
SelectCrayons("purple m", "tickle")

}
\references{
\url{http://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors}

\url{https://github.com/kbroman/broman/blob/master/R/brocolors.R}

\url{http://kbroman.org/blog/2017/05/02/crayon-colors-simplified/}
}
\seealso{
\code{\link{PlotCrayons}}, \code{\link{PlotColors}}
}
\keyword{utilities}
